Add lot in error message
Update to latest OCA conventions
