+++ stock_disallow_negative/__init__.py
@@ -1,3 +1,3 @@
-# -*- encoding: utf-8 -*-
+# -*- coding: utf-8 -*-
 
-from . import stock_disallow_negative
+from . import models
