+++ stock_disallow_negative/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import product
+from . import stock_quant
